feat(cli): add support for policy update from library
ALLY-849
